packpath: Add PackPath Python module
Reads a path and loads a `config.yaml` file from it to automatically
fill a `signalstickers-client` client, and submit stickers.

It subclasses signalstickers_client.models.LocalStickerPack to add a
`load_path()` method.

The module is meant for command line invocation, but can easily be used
in code:
  $ python3 -m packpath [path] --user [uuid_id] --password [password]

See `python3 -m packpath -h` for details on the YAML format, and how to
get the Signal user and password.